Static cast and warning
One more static cast, and a warning to my future self

+++ bytecode.cpp
@@ -1,13 +1,16 @@
 #include "bytecode.h"
 
 /**
+* NB: this one's pretty dangerous. Easy to reason through, but lots of casts
+* and bit manipulation.
+* 
 * @param constant : store this Value in the constant table
 * @param overflow : output param, index byte 2 - will be clobbered
 * @return : index (in constant table) byte 1
 */
 uint8_t Chunk::add_constant(Value constant, uint8_t& overflow) {
 	constants.push_back(constant);
-	uint16_t c = constants.size() - 1;
+	uint16_t c = static_cast<uint16_t>(constants.size() - 1);
 	overflow = (c > 225) ? static_cast<uint8_t>(c >> 8) : 0;
 	return (c > 255) ? static_cast<uint8_t>(c & 255) : static_cast<uint8_t>(c);
 };
